examples/avatar_agents/simli/requirements.txt
livekit
livekit-agents[openai,simli]
python-dotenv
httpx
